Move account datastore to new class.
All account handling in dictionary datastore and credentials file
handling moved to a new class.

The credentials service creates an instance of the datastore to
handle account storage.
