+++ mash/services/credentials/service.py
@@ -18,27 +18,14 @@
 
 import jwt
 import json
-import os
 
-from apscheduler import events
-from apscheduler.schedulers.background import BackgroundScheduler
-
-from contextlib import suppress
-from cryptography.fernet import Fernet
 from datetime import datetime, timedelta
-from pytz import utc
 
 # project
 from mash.services.mash_service import MashService
-from mash.services.credentials import (
-    add_account_to_db,
-    delete_account_from_db,
-    remove_account_from_groups
-)
-from mash.services.credentials.key_rotate import clean_old_keys, rotate_key
-from mash.services.jobcreator.accounts import accounts_template
+from mash.services.credentials import get_account_info
+from mash.services.credentials.account_datastore import AccountDatastore
 from mash.utils.json_format import JsonFormat
-from mash.mash_exceptions import MashCredentialsException
 
 
 class CredentialsService(MashService):
@@ -53,13 +40,12 @@ class CredentialsService(MashService):
         )
         self.accounts_file = self.config.get_accounts_file()
         self.credentials_directory = self.config.get_credentials_dir()
+        self.account_datastore = AccountDatastore(
+            self.accounts_file, self.credentials_directory,
+            self.encryption_keys_file, self._send_control_response
+        )
+
         self.jobs = {}
-
-        if not os.path.exists(self.encryption_keys_file):
-            self._create_encryption_keys_file()
-
-        if not os.path.exists(self.accounts_file):
-            self._write_accounts_to_file(accounts_template)
 
         self.bind_queue(
             self.service_exchange, self.add_account_key, self.listener_queue
@@ -68,12 +54,6 @@ class CredentialsService(MashService):
             self.service_exchange, self.delete_account_key, self.listener_queue
         )
         self._bind_credential_request_keys()
-
-        self.scheduler = BackgroundScheduler(timezone=utc)
-        self.scheduler.add_listener(
-            self._handle_key_rotation_result,
-            events.EVENT_JOB_EXECUTED | events.EVENT_JOB_ERROR
-        )
 
         self.restart_jobs(self._add_job)
         self.start()
@@ -92,10 +72,12 @@ class CredentialsService(MashService):
             self.jobs[job_id] = job_document
 
             if 'job_file' not in job_document:
-                testing_accounts = self._get_testing_accounts(
-                    job_document['cloud'], job_document['cloud_accounts'],
-                    job_document['requesting_user']
-                )
+                testing_accounts = \
+                    self.account_datastore.get_testing_accounts(
+                        job_document['cloud'], job_document['cloud_accounts'],
+                        job_document['requesting_user']
+                    )
+
                 job_document['cloud_accounts'] += testing_accounts
 
                 job_document['job_file'] = self.persist_job_config(
@@ -107,29 +89,6 @@ class CredentialsService(MashService):
                 job_id=job_id
             )
 
-    def _get_testing_accounts(self, cloud, cloud_accounts, requesting_user):
-        """
-        Return a list of testing accounts based on cloud accounts.
-
-        Only add an account to the list if it does not already exist.
-        """
-        accounts = self._get_accounts_from_file(cloud)
-        testing_accounts = []
-
-        for account in cloud_accounts:
-            info = self._get_account_info(
-                account, requesting_user, accounts
-            )
-
-            if info.get('testing_account') and \
-                    info['testing_account'] not in cloud_accounts and \
-                    info['testing_account'] not in testing_accounts:
-                testing_accounts.append(
-                    info['testing_account']
-                )
-
-        return testing_accounts
-
     def _bind_credential_request_keys(self):
         """
         Bind routing keys for all valid credential service requests.
@@ -138,74 +97,6 @@ class CredentialsService(MashService):
             self.bind_queue(
                 self.service_exchange, 'request.{0}'.format(service), 'request'
             )
-
-    def _check_credentials_exist(self, account, cloud, user):
-        """
-        Return True if the credentials file exists.
-        """
-        path = self._get_credentials_file_path(account, cloud, user)
-        return os.path.exists(path)
-
-    def _get_account_info(self, account, user, accounts):
-        """
-        Return info for the requested account.
-        """
-        try:
-            account_info = accounts['accounts'][user][account]
-        except KeyError:
-            raise MashCredentialsException(
-                'The requesting user {0}, does not have '
-                'the following account: {1}'.format(
-                    user, account
-                )
-            )
-
-        return account_info
-
-    def _check_job_accounts(
-        self, cloud, cloud_accounts,
-        cloud_groups, requesting_user, accounts
-    ):
-        """
-        Confirm all the accounts for the given user have credentials.
-        """
-        account_names = [account['name'] for account in cloud_accounts]
-        accounts_info = {}
-
-        for group in cloud_groups:
-            account_names += self._get_accounts_in_group(
-                group, requesting_user, accounts
-            )
-
-        for account in set(account_names):
-            info = self._get_account_info(
-                account, requesting_user, accounts
-            )
-            if info.get('testing_account'):
-                # If testing account does not exist raise exception
-                # and prevent job from entering queue.
-                self._get_account_info(
-                    info['testing_account'], requesting_user, accounts
-                )
-
-        for account in set(account_names):
-            accounts_info[account] = self._get_account_info(
-                account, requesting_user, accounts
-            )
-
-            exists = self._check_credentials_exist(
-                account, cloud, requesting_user
-            )
-
-            if not exists:
-                raise MashCredentialsException(
-                    'The requesting user {0}, does not have '
-                    'the following account: {1}'.format(
-                        requesting_user, account
-                    )
-                )
-
-        return accounts_info
 
     def _confirm_job(self, job_document):
         """
@@ -220,12 +111,9 @@ class CredentialsService(MashService):
         cloud_groups = job_document['cloud_groups']
         requesting_user = job_document['requesting_user']
 
-        accounts = self._get_accounts_from_file(cloud)
-
         try:
-            accounts_info = self._check_job_accounts(
-                cloud, cloud_accounts, cloud_groups, requesting_user,
-                accounts
+            accounts_info = self.account_datastore.check_job_accounts(
+                cloud, cloud_accounts, cloud_groups, requesting_user
             )
         except Exception as error:
             self._send_control_response(
@@ -252,14 +140,6 @@ class CredentialsService(MashService):
                 JsonFormat.json_message(job_response)
             )
 
-    def _create_encryption_keys_file(self):
-        """
-        Creates the keys file and stores a new key for use in encryption.
-        """
-        key = self._generate_encryption_key()
-        with open(self.encryption_keys_file, 'w') as keys_file:
-            keys_file.write(key)
-
     def _delete_job(self, job_id):
         """
         Remove job from dictionary.
@@ -280,55 +160,12 @@ class CredentialsService(MashService):
                 job_id=job_id
             )
 
-    def _generate_encryption_key(self):
-        """
-        Generates and returns a new Fernet key for encryption.
-        """
-        return Fernet.generate_key().decode()
-
-    def _get_accounts_from_file(self, cloud=None):
-        """
-        Return a dictionary of account information from accounts json file.
-        """
-        with open(self.accounts_file, 'r') as acnt_file:
-            accounts = json.load(acnt_file)
-
-        if cloud:
-            return accounts[cloud]
-        else:
-            return accounts
-
-    def _get_accounts_in_group(self, group, user, accounts_info):
-        """
-        Return a list of account names given the group name.
-        """
-        try:
-            accounts = accounts_info['groups'][user][group]
-        except KeyError:
-            raise MashCredentialsException(
-                'The requesting user: {0}, does not have the '
-                'following group: {1}'.format(
-                    user, group
-                )
-            )
-
-        return accounts
-
-    def _get_credentials_file_path(self, account, cloud, user):
-        """
-        Return the string path to the credentials file.
-
-        Based on user, cloud and account name.
-        """
-        path = os.path.join(
-            self.credentials_directory, user, cloud, account
-        )
-        return path
-
     def _get_credentials_response(self, job_id, issuer):
         """
         Return jwt encoded credentials response message.
         """
+        job = self.jobs[job_id]
+
         response = {
             'exp': datetime.utcnow() + timedelta(minutes=5),  # Expiration time
             'iat': datetime.utcnow(),  # Issued at time
@@ -336,21 +173,13 @@ class CredentialsService(MashService):
             'iss': self.service_exchange,  # Issuer
             'aud': issuer,  # audience
             'id': job_id,
-            'credentials': self._retrieve_credentials(job_id)
+            'credentials': self.account_datastore.retrieve_credentials(
+                job['cloud_accounts'], job['cloud'], job['requesting_user']
+            )
         }
         return jwt.encode(
             response, self.jwt_secret, algorithm=self.jwt_algorithm
         )
-
-    def _get_encrypted_credentials(self, account, cloud, user):
-        """
-        Return encrypted credentials string from file.
-        """
-        path = self._get_credentials_file_path(account, cloud, user)
-        with open(path, 'r') as credentials_file:
-            credentials = credentials_file.read()
-
-        return credentials.strip()
 
     def _handle_account_request(self, message):
         """
@@ -443,26 +272,6 @@ class CredentialsService(MashService):
 
         message.ack()
 
-    def _handle_key_rotation_result(self, event):
-        """
-        Callback when key rotation cron finishes.
-
-        If the rotation does not finish successfully the old key
-        is left in key file.
-
-        Once a successful rotation happens all old keys are purged.
-        """
-        if event.exception:
-            self.log.error(
-                'Key rotation did not finish successfully.'
-                ' Old key will remain in key file.'
-            )
-        else:
-            clean_old_keys(
-                self.encryption_keys_file, self._send_control_response
-            )
-            self.log.info('Key rotation finished.')
-
     def _publish_credentials_response(self, credentials_response, issuer):
         """
         Publish the encoded JWT with secrets to the calling service.
@@ -473,35 +282,6 @@ class CredentialsService(MashService):
         self._publish(
             issuer, self.credentials_response_key, credentials_response
         )
-
-    def _remove_credentials_file(self, account_name, cloud, user):
-        """
-        Attempt to remove the credentials file for account.
-        """
-        self.log.info(
-            'Deleting credentials for account: '
-            '{0}, cloud: {1}, user: {2}.'.format(
-                account_name, cloud, user
-            )
-        )
-
-        path = self._get_credentials_file_path(account_name, cloud, user)
-
-        with suppress(Exception):
-            os.remove(path)
-
-    def _retrieve_credentials(self, job_id):
-        """
-        Retrieve the encrypted credentials strings for the requested job_id.
-        """
-        job = self.jobs[job_id]
-
-        credentials = {}
-        for account in job['cloud_accounts']:
-            credentials[account] = self._get_encrypted_credentials(
-                account, job['cloud'], job['requesting_user']
-            )
-        return credentials
 
     def _send_control_response(self, message, success=True, job_id=None):
         job_metadata = {}
@@ -526,7 +306,9 @@ class CredentialsService(MashService):
             credentials_response = self._get_credentials_response(
                 job_id, issuer
             )
-            message = JsonFormat.json_message({'jwt_token': credentials_response.decode()})
+            message = JsonFormat.json_message(
+                {'jwt_token': credentials_response.decode()}
+            )
 
             self.log.info(
                 'Received credentials request from {0} for job: {1}.'.format(
@@ -541,67 +323,6 @@ class CredentialsService(MashService):
                 success=False, job_id=job_id
             )
 
-    def _start_rotation_job(self):
-        """
-        Schedule new key rotation cron job in background scheduler.
-
-        Job is run once a month on the first Saturday at 0000.
-        """
-        self.scheduler.add_job(
-            rotate_key,
-            'cron',
-            args=(
-                self.credentials_directory,
-                self.encryption_keys_file,
-                self._send_control_response
-            ),
-            day='1st sat,3rd sat',
-            hour='0',
-            minute='0'
-        )
-
-    def _store_encrypted_credentials(
-        self, account, credentials, cloud, user
-    ):
-        """
-        Store the provided credentials encrypted on disk.
-
-        Expected credentials as a json string.
-
-        Example: {"access_key_id": "key123", "secret_access_key": "123456"}
-
-        Path is based on the user, cloud and account.
-        """
-        self.log.info(
-            'Storing credentials for account: '
-            '{0}, cloud: {1}, user: {2}.'.format(
-                account, cloud, user
-            )
-        )
-
-        path = self._get_credentials_file_path(account, cloud, user)
-
-        credentials_dir = os.path.dirname(path)
-        if not os.path.isdir(credentials_dir):
-            os.makedirs(credentials_dir)
-
-        try:
-            with open(path, 'w') as creds_file:
-                creds_file.write(credentials)
-        except Exception as error:
-            self.log.error(
-                'Unable to store credentials: {0}.'.format(error)
-            )
-
-    def _write_accounts_to_file(self, accounts):
-        """
-        Update accounts file with provided accounts dictionary.
-        """
-        account_info = JsonFormat.json_message(accounts)
-
-        with open(self.accounts_file, 'w') as account_file:
-            account_file.write(account_info)
-
     def add_account(self, message):
         """
         Add new cloud account to MASH.
@@ -612,28 +333,23 @@ class CredentialsService(MashService):
             )
         )
 
-        cloud = message['cloud']
-        account_name = message['account_name']
-        requesting_user = message['requesting_user']
-
-        accounts = self._get_accounts_from_file()
-
         try:
-            accounts = add_account_to_db(message, accounts)
+            account_info = get_account_info(message)
         except Exception as error:
             self.log.warning(
                 'Failed to add account to database: {0}'.format(error)
             )
             return
 
-        self._write_accounts_to_file(accounts)
+        cloud = message['cloud']
+        credentials = message['credentials']
+        account_name = message['account_name']
+        requesting_user = message['requesting_user']
+        group_name = message.get('group')
 
-        credentials = self.encrypt_credentials(
-            json.dumps(message['credentials'])
-        )
-
-        self._store_encrypted_credentials(
-            account_name, credentials, cloud, requesting_user
+        self.account_datastore.add_account(
+            account_info, cloud, account_name, requesting_user, credentials,
+            group_name
         )
 
     def delete_account(self, message):
@@ -644,35 +360,14 @@ class CredentialsService(MashService):
         cloud = message['cloud']
         requesting_user = message['requesting_user']
 
-        self._remove_credentials_file(
-            account_name, cloud, requesting_user
+        self.account_datastore.delete_account(
+            requesting_user, account_name, cloud
         )
-
-        accounts = self._get_accounts_from_file()
-
-        try:
-            accounts = delete_account_from_db(
-                accounts, requesting_user, account_name, cloud
-            )
-        except Exception:
-            self.log.error(
-                'Account {0} does not exist for {1}.'.format(
-                    account_name, requesting_user
-                )
-            )
-        else:
-            accounts = remove_account_from_groups(
-                accounts, account_name, cloud, requesting_user
-            )
-            self._write_accounts_to_file(accounts)
 
     def start(self):
         """
         Start credentials service.
         """
-        self.scheduler.start()
-        self._start_rotation_job()
-
         self.consume_queue(self._handle_job_documents)
         self.consume_queue(
             self._handle_account_request, queue_name=self.listener_queue
@@ -688,5 +383,5 @@ class CredentialsService(MashService):
         except Exception:
             raise
         finally:
-            self.scheduler.shutdown()
+            self.account_datastore.shutdown()
             self.close_connection()
