+++ test/unit/services/credentials/service_test.py
@@ -1,4 +1,3 @@
-import io
 import jwt
 import json
 
@@ -7,8 +6,7 @@ from unittest.mock import call, MagicMock, Mock, patch
 
 from mash.services.mash_service import MashService
 from mash.services.credentials.service import CredentialsService
-from mash.services.credentials.key_rotate import rotate_key
-from mash.services.jobcreator.accounts import accounts_template
+from mash.services.credentials.account_datastore import AccountDatastore
 from mash.utils.json_format import JsonFormat
 from mash.mash_exceptions import MashCredentialsException
 
@@ -32,6 +30,11 @@ class TestCredentialsService(object):
         self.service.credentials_directory = '/var/lib/mash/credentials/'
         self.service.accounts_file = '../data/accounts.json'
         self.service.encryption_keys_file = '/var/lib/mash/encryption_keys'
+        self.service.account_datastore = AccountDatastore(
+            self.service.accounts_file, self.service.credentials_directory,
+            self.service.encryption_keys_file,
+            self.service._send_control_response
+        )
 
         self.service.channel = Mock()
         self.service.channel.basic_ack.return_value = None
@@ -39,20 +42,16 @@ class TestCredentialsService(object):
         self.service.jobs = {}
         self.service.log = Mock()
 
-    @patch.object(CredentialsService, '_write_accounts_to_file')
-    @patch.object(CredentialsService, '_create_encryption_keys_file')
-    @patch('mash.services.credentials.service.os')
+    @patch('mash.services.credentials.service.AccountDatastore')
     @patch.object(CredentialsService, 'set_logfile')
     @patch.object(CredentialsService, 'start')
     @patch.object(CredentialsService, '_bind_credential_request_keys')
     @patch.object(CredentialsService, 'restart_jobs')
     def test_post_init(
         self, mock_restart_jobs, mock_bind_cred_req_keys, mock_start,
-        mock_set_logfile, mock_os, mock_create_keys_file,
-        mock_write_accounts_to_file
+        mock_set_logfile, mock_datastore
     ):
         self.service.config = self.config
-        mock_os.path.exists.return_value = False
         self.config.get_log_file.return_value = \
             '/var/log/mash/credentials_service.log'
 
@@ -66,29 +65,19 @@ class TestCredentialsService(object):
             '/var/log/mash/credentials_service.log'
         )
 
-        mock_create_keys_file.assert_called_once_with()
-        mock_write_accounts_to_file.assert_called_once_with(
-            accounts_template
-        )
         mock_bind_cred_req_keys.assert_called_once_with()
         mock_restart_jobs.assert_called_once_with(self.service._add_job)
         mock_start.assert_called_once_with()
 
-    @patch.object(CredentialsService, '_get_accounts_from_file')
+    @patch.object(AccountDatastore, 'get_testing_accounts')
     @patch.object(CredentialsService, 'persist_job_config')
     @patch.object(CredentialsService, '_send_control_response')
     def test_credentials_add_job(
         self, mock_send_control_response, mock_persist_job_config,
-        mock_get_accounts_from_file
+        mock_get_testing_accounts
     ):
         mock_persist_job_config.return_value = 'temp-config.json'
-        mock_get_accounts_from_file.return_value = {
-            'accounts': {
-                'user1': {
-                    'test-gce': {'testing_account': 'tester'}
-                }
-            }
-        }
+        mock_get_testing_accounts.return_value = ['tester']
 
         self.service._add_job({
             'id': '1',
@@ -137,76 +126,12 @@ class TestCredentialsService(object):
             call('credentials', 'request.pint', 'request')
         ])
 
-    @patch('mash.services.credentials.service.os.path')
-    def test_check_credentials_exist(self, mock_path):
-        mock_path.exists.return_value = True
-
-        assert self.service._check_credentials_exist(
-            'account1', 'ec2', 'user1'
-        )
-
-    def test_get_account_info_error(self):
-        accounts = {
-            'accounts': {},
-            'groups': {}
-        }
-
-        with raises(MashCredentialsException) as error:
-            self.service._get_account_info(
-                'test-aws', 'user1', accounts
-            )
-
-        msg = 'The requesting user user1, does not have ' \
-              'the following account: test-aws'
-        assert str(error.value) == msg
-
-    @patch.object(CredentialsService, '_check_credentials_exist')
-    def test_check_job_accounts(
-        self, mock_check_credentials_exist
-    ):
-        mock_check_credentials_exist.return_value = True
-
-        accounts = {
-            'accounts': {
-                'user1': {
-                    'test-aws': {
-                        'testing_account': 'tester'
-                    },
-                    'tester': {}
-                }
-            },
-            'groups': {
-                'user1': {
-                    'test': ['test-aws']
-                }
-            }
-        }
-
-        self.service._check_job_accounts(
-            'ec2', [{'name': 'test-aws'}], ['test'], 'user1',
-            accounts
-        )
-
-        # Account does not exist for user
-        mock_check_credentials_exist.return_value = False
-
-        with raises(MashCredentialsException) as error:
-            self.service._check_job_accounts(
-                'ec2', [{'name': 'test-aws'}], ['test'], 'user1',
-                accounts
-            )
-
-        msg = 'The requesting user user1, does not have ' \
-              'the following account: test-aws'
-        assert str(error.value) == msg
-
-    @patch.object(CredentialsService, '_get_accounts_from_file')
-    @patch.object(CredentialsService, '_check_job_accounts')
+    @patch.object(AccountDatastore, 'check_job_accounts')
     @patch.object(CredentialsService, '_send_control_response')
     @patch.object(CredentialsService, '_publish')
     def test_confirm_job(
         self, mock_publish, mock_send_control_response,
-        mock_check_job_accounts, mock_get_accounts_from_file
+        mock_check_job_accounts
     ):
         mock_check_job_accounts.return_value = {'accounts': 'info'}
 
@@ -250,18 +175,6 @@ class TestCredentialsService(object):
             )
         )
 
-    @patch.object(CredentialsService, '_generate_encryption_key')
-    def test_create_encryption_keys_file(self, mock_generate_encryption_key):
-        mock_generate_encryption_key.return_value = '1234567890'
-
-        with patch('builtins.open', create=True) as mock_open:
-            mock_open.return_value = MagicMock(spec=io.IOBase)
-            self.service._create_encryption_keys_file()
-            file_handle = mock_open.return_value.__enter__.return_value
-            file_handle.write.assert_called_with(
-                u'1234567890'
-            )
-
     @patch.object(CredentialsService, 'remove_file')
     @patch.object(CredentialsService, '_send_control_response')
     def test_credentials_delete_job(
@@ -285,68 +198,6 @@ class TestCredentialsService(object):
             'Job deletion failed, job is not queued.',
             success=False, job_id='1'
         )
-
-    @patch('mash.services.credentials.service.Fernet')
-    def test_generate_encryption_key(self, mock_fernet):
-        mock_fernet.generate_key.return_value = b'1234567890'
-        value = self.service._generate_encryption_key()
-
-        assert value == '1234567890'
-
-    def test_get_accounts_from_file(self):
-        with patch('builtins.open', create=True) as mock_open:
-            mock_open.return_value = MagicMock(spec=io.IOBase)
-            file_handle = mock_open.return_value.__enter__.return_value
-            file_handle.read.return_value = '{"ec2": {"account": "info"}}'
-            result = self.service._get_accounts_from_file('ec2')
-
-        assert result['account'] == 'info'
-
-        # Test get accounts with no cloud
-        with patch('builtins.open', create=True) as mock_open:
-            mock_open.return_value = MagicMock(spec=io.IOBase)
-            file_handle = mock_open.return_value.__enter__.return_value
-            file_handle.read.return_value = '{"ec2": {"account": "info"}}'
-            result = self.service._get_accounts_from_file()
-
-        assert result['ec2']['account'] == 'info'
-
-    def test_get_accounts_in_group(self):
-        accounts_info = {
-            'groups': {
-                'user1': {'test': ['test-1', 'test-2']}
-            }
-        }
-        accounts = self.service._get_accounts_in_group(
-            'test', 'user1', accounts_info
-        )
-        assert accounts == ['test-1', 'test-2']
-
-    def test_get_accounts_in_group_error(self):
-        accounts_info = {
-            'groups': {}
-        }
-
-        with raises(MashCredentialsException) as error:
-            self.service._get_accounts_in_group(
-                'test', 'user1', accounts_info
-            )
-
-        msg = 'The requesting user: user1, does ' \
-              'not have the following group: test'
-        assert str(error.value) == msg
-
-    def test_get_encrypted_credentials(self):
-        with patch('builtins.open', create=True) as mock_open:
-            mock_open.return_value = MagicMock(spec=io.IOBase)
-            file_handle = mock_open.return_value.__enter__.return_value
-            file_handle.read.return_value = 'secret_stuff'
-
-            result = self.service._get_encrypted_credentials(
-                'account1', 'ec2', 'user1'
-            )
-
-        assert result == 'secret_stuff'
 
     @patch.object(CredentialsService, 'add_account')
     @patch.object(CredentialsService, '_send_control_response')
@@ -478,64 +329,17 @@ class TestCredentialsService(object):
         )
         message.ack.assert_called_once_with()
 
-    @patch('mash.services.credentials.service.clean_old_keys')
-    def test_handle_key_rotation_result(self, mock_clean_old_keys):
-        # Test exception case
-        event = MagicMock()
-        event.exception = 'Broken!'
-
-        self.service._handle_key_rotation_result(event)
-        self.service.log.error.assert_called_once_with(
-            'Key rotation did not finish successfully.'
-            ' Old key will remain in key file.'
-        )
-
-        # Test success
-        event.exception = None
-
-        self.service._handle_key_rotation_result(event)
-        mock_clean_old_keys.assert_called_once_with(
-            '/var/lib/mash/encryption_keys',
-            self.service._send_control_response
-        )
-        self.service.log.info.assert_called_once_with(
-            'Key rotation finished.'
-        )
-
-    @patch.object(CredentialsService, 'encrypt_credentials')
-    @patch('mash.services.credentials.service.os.makedirs')
-    @patch('mash.services.credentials.service.os.path.isdir')
-    def test_store_encrypted_credentials(
-        self, mock_isdir, mock_makedirs, mock_encrypt_creds
-    ):
-        mock_isdir.return_value = False
-
-        with patch('builtins.open', create=True) as mock_open:
-            mock_open.return_value = MagicMock(spec=io.IOBase)
-            self.service._store_encrypted_credentials(
-                'account1', 'encrypted_secrets', 'ec2', 'user1'
-            )
-            file_handle = mock_open.return_value.__enter__.return_value
-            file_handle.write.assert_called_once_with('encrypted_secrets')
-            self.service.log.info.assert_called_once_with(
-                'Storing credentials for account: '
-                'account1, cloud: ec2, user: user1.'
-            )
-
-        with patch('builtins.open', create=True) as mock_open:
-            mock_open.return_value = MagicMock(spec=io.IOBase)
-            file_handle = mock_open.return_value.__enter__.return_value
-            file_handle.write.side_effect = Exception('Cannot write file')
-
-            self.service._store_encrypted_credentials(
-                'account1', 'encrypted_secrets', 'ec2', 'user1'
-            )
-            self.service.log.error.assert_called_once_with(
-                'Unable to store credentials: Cannot write file.'
-            )
-
-    @patch.object(CredentialsService, '_retrieve_credentials')
+    @patch.object(AccountDatastore, 'retrieve_credentials')
     def test_get_credentials_response(self, mock_retrieve_credentials):
+        job = {
+            'id': '1',
+            'cloud': 'ec2',
+            'job_file': 'temp-config.json',
+            'cloud_accounts': ['test-aws'],
+            'requesting_user': 'user1'
+        }
+
+        self.service.jobs['1'] = job
         self.service.jwt_algorithm = 'HS256'
         self.service.jwt_secret = 'super.secret'
 
@@ -548,7 +352,9 @@ class TestCredentialsService(object):
             issuer='credentials', audience='testing'
         )
 
-        mock_retrieve_credentials.assert_called_once_with('1')
+        mock_retrieve_credentials.assert_called_once_with(
+            ['test-aws'], 'ec2', 'user1'
+        )
         assert payload['id'] == '1'
         assert payload['sub'] == 'credentials_response'
 
@@ -610,60 +416,8 @@ class TestCredentialsService(object):
             extra={}
         )
 
-    @patch.object(CredentialsService, '_get_encrypted_credentials')
-    def test_retrieve_credentials(self, mock_get_encrypt_creds):
-        mock_get_encrypt_creds.return_value = 'encrypted_string'
-        self.service.jobs = {
-            '1': {
-                'id': '1',
-                'cloud': 'ec2',
-                'cloud_accounts': ['test-aws'],
-                'requesting_user': 'user1',
-                'last_service': 'pint'
-            }
-        }
-        credentials = self.service._retrieve_credentials('1')
-
-        assert credentials['test-aws'] == 'encrypted_string'
-
-    def test_credentials_start_rotation_job(self):
-        scheduler = Mock()
-        self.service.scheduler = scheduler
-
-        self.service._start_rotation_job()
-
-        self.service.scheduler.add_job.assert_called_once_with(
-            rotate_key,
-            'cron',
-            args=(
-                '/var/lib/mash/credentials/',
-                '/var/lib/mash/encryption_keys',
-                self.service._send_control_response
-            ),
-            day='1st sat,3rd sat',
-            hour='0',
-            minute='0'
-        )
-
-    def test_credentials_write_accounts_to_file(self):
-        accounts = {'test': 'accounts'}
-
-        with patch('builtins.open', create=True) as mock_open:
-            mock_open.return_value = MagicMock(spec=io.IOBase)
-            self.service._write_accounts_to_file(accounts)
-            file_handle = mock_open.return_value.__enter__.return_value
-            file_handle.write.assert_called_with(
-                u'{\n    "test": "accounts"\n}'
-            )
-
-    @patch.object(CredentialsService, '_store_encrypted_credentials')
-    @patch.object(CredentialsService, '_write_accounts_to_file')
-    @patch.object(CredentialsService, 'encrypt_credentials')
-    @patch.object(CredentialsService, '_get_accounts_from_file')
-    def test_credentials_add_account_ec2(
-        self, mock_get_acnts_from_file, mock_encrypt_creds,
-        mock_write_accounts_to_file, mock_store_encrypted_creds
-    ):
+    @patch.object(AccountDatastore, 'add_account')
+    def test_credentials_add_account_ec2(self, mock_add_account):
         message = {
             'account_name': 'acnt123',
             'credentials': {'creds': 'data'},
@@ -674,33 +428,11 @@ class TestCredentialsService(object):
             'group': 'group123'
         }
 
-        with open(self.service.accounts_file) as f:
-            accounts = json.load(f)
-        mock_get_acnts_from_file.return_value = accounts
-        mock_encrypt_creds.return_value = 'encryptedcreds'
-
         self.service.add_account(message)
+        assert mock_add_account.call_count == 1
 
-        self.service._write_accounts_to_file.assert_called_once_with(accounts)
-        self.service._store_encrypted_credentials.assert_called_once_with(
-            'acnt123', 'encryptedcreds', 'ec2', 'user1'
-        )
-
-        assert accounts['ec2']['accounts']['user1']['acnt123'] == {
-            'additional_regions': None,
-            'partition': 'aws',
-            'region': 'us-east-1'
-        }
-        assert accounts['ec2']['groups']['user1']['group123'] == ['acnt123']
-
-    @patch.object(CredentialsService, '_store_encrypted_credentials')
-    @patch.object(CredentialsService, '_write_accounts_to_file')
-    @patch.object(CredentialsService, 'encrypt_credentials')
-    @patch.object(CredentialsService, '_get_accounts_from_file')
-    def test_credentials_add_account_azure(
-        self, mock_get_acnts_from_file, mock_encrypt_creds,
-        mock_write_accounts_to_file, mock_store_encrypted_creds
-    ):
+    @patch.object(AccountDatastore, 'add_account')
+    def test_credentials_add_account_azure(self, mock_add_account):
         message = {
             "account_name": "test-azure",
             "credentials": {"encrypted": "creds"},
@@ -715,59 +447,22 @@ class TestCredentialsService(object):
             "destination_storage_account": "sa_2"
         }
 
-        with open(self.service.accounts_file) as f:
-            accounts = json.load(f)
-        mock_get_acnts_from_file.return_value = accounts
-        mock_encrypt_creds.return_value = 'encryptedcreds'
-
         self.service.add_account(message)
+        assert mock_add_account.call_count == 1
 
-        self.service._write_accounts_to_file.assert_called_once_with(accounts)
-        self.service._store_encrypted_credentials.assert_called_once_with(
-            'test-azure', 'encryptedcreds', 'azure', 'user1'
-        )
-
-        account = accounts['azure']['accounts']['user1']['test-azure']
-        assert account['region'] == 'southcentralus'
-        assert account['source_container'] == 'container1'
-        assert account['source_resource_group'] == 'rg_1'
-        assert account['source_storage_account'] == 'sa_1'
-        assert account['destination_container'] == 'container2'
-        assert account['destination_resource_group'] == 'rg_2'
-        assert account['destination_storage_account'] == 'sa_2'
-
-    @patch.object(CredentialsService, '_store_encrypted_credentials')
-    @patch.object(CredentialsService, '_write_accounts_to_file')
-    @patch.object(CredentialsService, 'encrypt_credentials')
-    @patch.object(CredentialsService, '_get_accounts_from_file')
-    def test_credentials_add_account_gce(
-        self, mock_get_acnts_from_file, mock_encrypt_creds,
-        mock_write_accounts_to_file, mock_store_encrypted_creds
-    ):
+    @patch.object(AccountDatastore, 'add_account')
+    def test_credentials_add_account_gce(self, mock_add_account):
         message = {
-            "account_name": "test-gce",
-            "bucket": "images",
-            "credentials": {"encrypted": "creds"},
-            "cloud": "gce",
-            "region": "us-west2",
-            "requesting_user": "user1"
+            'account_name': 'test-gce',
+            'bucket': 'images',
+            'credentials': {'encrypted': 'creds'},
+            'cloud': 'gce',
+            'region': 'us-west2',
+            'requesting_user': 'user1'
         }
 
-        with open(self.service.accounts_file) as f:
-            accounts = json.load(f)
-        mock_get_acnts_from_file.return_value = accounts
-        mock_encrypt_creds.return_value = 'encryptedcreds'
-
         self.service.add_account(message)
-
-        self.service._write_accounts_to_file.assert_called_once_with(accounts)
-        self.service._store_encrypted_credentials.assert_called_once_with(
-            'test-gce', 'encryptedcreds', 'gce', 'user1'
-        )
-
-        account = accounts['gce']['accounts']['user1']['test-gce']
-        assert account['region'] == 'us-west2'
-        assert account['bucket'] == 'images'
+        assert mock_add_account.call_count == 1
 
     def test_credentials_add_account_invalid(self):
         message = {
@@ -784,127 +479,29 @@ class TestCredentialsService(object):
             'Failed to add account to database: CSP fake is not supported.'
         )
 
-    @patch.object(CredentialsService, '_get_credentials_file_path')
-    @patch('mash.services.credentials.service.os')
-    @patch.object(CredentialsService, '_write_accounts_to_file')
-    @patch.object(CredentialsService, '_get_accounts_from_file')
-    def test_credentials_delete_account(
-        self, mock_get_acnts_from_file, mock_write_accounts_to_file,
-        mock_os, mock_get_cred_path
-    ):
-        mock_get_cred_path.return_value = \
-            '/var/lib/mash/credentials/user2/ec2/test-aws'
-
-        with open(self.service.accounts_file) as f:
-            accounts = json.load(f)
-        mock_get_acnts_from_file.return_value = accounts
-
+    @patch.object(AccountDatastore, 'delete_account')
+    def test_credentials_delete_account(self, mock_delete_account):
         message = {
-            "account_name": "test-aws",
-            "cloud": "ec2",
-            "requesting_user": "user2"
+            'account_name': 'test-aws',
+            'cloud': 'ec2',
+            'requesting_user': 'user2'
         }
 
         self.service.delete_account(message)
 
-        mock_get_acnts_from_file.assert_called_once_with()
-        self.service.log.info.assert_called_once_with(
-            'Deleting credentials for account: '
-            'test-aws, cloud: ec2, user: user2.'
-        )
-        mock_os.remove.assert_called_once_with(
-            '/var/lib/mash/credentials/user2/ec2/test-aws'
-        )
-        mock_write_accounts_to_file.assert_called_once_with(
-            {
-                'ec2': {
-                    'accounts': {
-                        'user2': {
-                            'test-aws-gov': {
-                                'additional_regions': [
-                                    {
-                                        'name': 'ap-northeast-4',
-                                        'helper_image': 'ami-82444aff'
-                                    }
-                                ],
-                                'partition': 'aws-us-gov'
-                            }
-                        }
-                    },
-                    'groups': {
-                        'user1': {
-                            'test1': []
-                        },
-                        'user2': {
-                            'test': ['test-aws-gov']
-                        }
-                    }
-                },
-                "azure": {
-                    "groups": {},
-                    "accounts": {
-                        "user1": {
-                            "test123": {
-                                "container_name": "container1",
-                                "region": "centralus",
-                                "resource_group": "rg_123",
-                                "storage_account": "sa_123"
-                            }
-                        }
-                    }
-                },
-                "gce": {
-                    "groups": {},
-                    "accounts": {
-                        "user1": {
-                            "test123": {
-                                "bucket": "images",
-                                "region": "us-west2"
-                            }
-                        }
-                    }
-                }
-            }
+        mock_delete_account.assert_called_once_with(
+            'user2', 'test-aws', 'ec2'
         )
 
-    @patch.object(CredentialsService, '_remove_credentials_file')
-    @patch.object(CredentialsService, '_get_accounts_from_file')
-    def test_credentials_delete_account_invalid(
-        self, mock_get_acnts_from_file, mock_rmv_creds_file
-    ):
-        with open(self.service.accounts_file) as f:
-            accounts = json.load(f)
-        mock_get_acnts_from_file.return_value = accounts
-
-        message = {
-            "account_name": "fake",
-            "cloud": "ec2",
-            "requesting_user": "user2"
-        }
-
-        self.service.delete_account(message)
-        mock_rmv_creds_file.assert_called_once_with(
-            'fake', 'ec2', 'user2'
-        )
-        self.service.log.error.assert_called_once_with(
-            'Account fake does not exist for user2.'
-        )
-
-    @patch.object(CredentialsService, '_start_rotation_job')
     @patch.object(CredentialsService, 'consume_credentials_queue')
     @patch.object(CredentialsService, 'consume_queue')
     @patch.object(CredentialsService, 'close_connection')
     def test_credentials_start(
         self, mock_close_connection, mock_consume_queue,
-        mock_consume_creds_queue, mock_start_rotation_job
+        mock_consume_creds_queue
     ):
-        scheduler = Mock()
-        self.service.scheduler = scheduler
-
         self.service.start()
 
-        scheduler.start.assert_called_once_with()
-        mock_start_rotation_job.assert_called_once_with()
         self.service.channel.start_consuming.assert_called_once_with()
         mock_consume_queue.has_calls([
             call(self.service._handle_job_documents),
@@ -915,26 +512,25 @@ class TestCredentialsService(object):
         )
         mock_close_connection.assert_called_once_with()
 
+    @patch.object(AccountDatastore, 'shutdown')
     @patch.object(CredentialsService, 'consume_credentials_queue')
     @patch.object(CredentialsService, 'consume_queue')
     @patch.object(CredentialsService, 'close_connection')
     def test_credentials_start_exception(
         self, mock_close_connection, mock_consume_queue,
-        mock_consume_creds_queue
+        mock_consume_creds_queue, mock_datastore_shutdown
     ):
-        scheduler = Mock()
-        self.service.scheduler = scheduler
-
         self.service.channel.start_consuming.side_effect = KeyboardInterrupt()
         self.service.start()
 
         mock_close_connection.assert_called_once_with()
+        mock_datastore_shutdown.assert_called_once_with()
         mock_close_connection.reset_mock()
         self.service.channel.start_consuming.side_effect = Exception(
-            'Cannot start scheduler.'
+            'Cannot start.'
         )
 
         with raises(Exception) as error:
             self.service.start()
 
-        assert 'Cannot start scheduler.' == str(error.value)
+        assert 'Cannot start.' == str(error.value)
